relocate test to appropriate file
