+++ coldfront/plugins/qumulo/tests/test_tasks.py
@@ -5,16 +5,25 @@ from unittest.mock import patch, MagicMock
 from coldfront.plugins.qumulo.tests.utils.mock_data import (
     build_models,
     create_allocation,
+    mock_quotas,
+    quota_data_coldfront_allocations,
 )
 from coldfront.plugins.qumulo.tasks import (
     poll_ad_group,
     poll_ad_groups,
     conditionally_update_storage_allocation_status,
     conditionally_update_storage_allocation_statuses,
+    ingest_quotas_with_daily_usage,
 )
 from coldfront.plugins.qumulo.utils.acl_allocations import AclAllocations
 
-from coldfront.core.allocation.models import Allocation, AllocationStatusChoice
+from coldfront.core.allocation.models import (
+    Allocation,
+    AllocationStatusChoice,
+    AllocationAttribute,
+    AllocationAttributeType,
+)
+
 from coldfront.core.resource.models import Resource
 
 from qumulo.lib.request import RequestError
@@ -136,7 +145,9 @@ class TestPollAdGroups(TestCase):
         )
         acl_allocation_c.resources.add(resource_b)
 
-        with patch("coldfront.plugins.qumulo.tasks.poll_ad_group") as poll_ad_group_mock:
+        with patch(
+            "coldfront.plugins.qumulo.tasks.poll_ad_group"
+        ) as poll_ad_group_mock:
             poll_ad_groups()
 
             self.assertEqual(poll_ad_group_mock.call_count, 2)
@@ -244,4 +255,127 @@ class TestStorageAllocationStatuses(TestCase):
 
             self.assertEqual(
                 conditionally_update_storage_allocation_status_mock.call_count, 2
+            )
+
+
+class TestIngestQuotasWithDailyUsages(TestCase):
+    def setUp(self) -> None:
+
+        self.client = Client()
+        build_data = build_models()
+
+        self.project = build_data["project"]
+        self.user = build_data["user"]
+        self.quotas = mock_quotas
+
+        for index, (path, details) in enumerate(
+            quota_data_coldfront_allocations.items()
+        ):
+
+            form_data = {
+                "storage_filesystem_path": path,
+                "storage_export_path": path,
+                "storage_name": f"for_tester_{index}",
+                "storage_quota": details.get("limit"),
+                "protocols": ["nfs"],
+                "rw_users": [f"user_{index}_rw"],
+                "ro_users": [f"user_{index}_ro"],
+                "storage_ticket": f"ITSD-{index}",
+                "cost_center": "Uncle Pennybags",
+                "department_number": "Time Travel Services",
+                "service_rate": "general",
+            }
+            create_allocation(project=self.project, user=self.user, form_data=form_data)
+
+        self.storage_filesystem_path_attribute_type = (
+            AllocationAttributeType.objects.get(name="storage_filesystem_path")
+        )
+        self.storage_quota_attribute_type = AllocationAttributeType.objects.get(
+            name="storage_quota"
+        )
+
+        return super().setUp()
+
+    def test_after_allocation_create_usage_is_zero(self) -> None:
+
+        # after allocations are created, expect usage to be zero
+        for path in quota_data_coldfront_allocations:
+            allocation_attribute_usage = None
+            try:
+                storage_filesystem_path_attribute = AllocationAttribute.objects.get(
+                    value=path,
+                    allocation_attribute_type=self.storage_filesystem_path_attribute_type,
+                )
+                allocation = storage_filesystem_path_attribute.allocation
+                storage_quota_attribute_type = AllocationAttribute.objects.get(
+                    allocation=allocation,
+                    allocation_attribute_type=self.storage_quota_attribute_type,
+                )
+                allocation_attribute_usage = (
+                    storage_quota_attribute_type.allocationattributeusage
+                )
+            except AllocationAttribute.DoesNotExist:
+                # When the storage_path_attribute for path is not found,
+                # the allocation_attribute_usage should not exist.
+                self.assertIsNone(allocation_attribute_usage)
+                continue
+
+            self.assertEqual(allocation_attribute_usage.value, 0)
+            self.assertEqual(allocation_attribute_usage.history.first().value, 0)
+
+    @patch("coldfront.plugins.qumulo.tasks.QumuloAPI")
+    def test_after_getting_daily_usages_from_qumulo_api(
+        self, qumulo_api_mock: MagicMock
+    ) -> None:
+        qumulo_api = MagicMock()
+        qumulo_api.get_all_quotas_with_usage.return_value = mock_quotas
+        qumulo_api_mock.return_value = qumulo_api
+
+        exceptionRaised = False
+        try:
+            ingest_quotas_with_daily_usage()
+        except:
+            exceptionRaised = True
+
+        self.assertFalse(exceptionRaised)
+
+        for qumulo_quota in self.quotas["quotas"]:
+
+            allocation_attribute_usage = None
+            try:
+                try:
+                    storage_filesystem_path_attribute = AllocationAttribute.objects.get(
+                        value=qumulo_quota.get("path"),
+                        allocation_attribute_type=self.storage_filesystem_path_attribute_type,
+                    )
+                except AllocationAttribute.DoesNotExist:
+                    path = qumulo_quota.get("path")
+                    if path[-1] != "/":
+                        continue
+
+                    storage_filesystem_path_attribute = AllocationAttribute.objects.get(
+                        value=path[:-1],
+                        allocation_attribute_type=self.storage_filesystem_path_attribute_type,
+                    )
+
+                allocation = storage_filesystem_path_attribute.allocation
+                storage_quota_attribute = AllocationAttribute.objects.get(
+                    allocation=allocation,
+                    allocation_attribute_type=self.storage_quota_attribute_type,
+                )
+
+                allocation_attribute_usage = (
+                    storage_quota_attribute.allocationattributeusage
+                )
+            except AllocationAttribute.DoesNotExist:
+                # When the storage_path_attribute for path is not found,
+                # the allocation_attribute_usage should not exist.
+                self.assertIsNone(allocation_attribute_usage)
+                continue
+
+            usage = int(qumulo_quota.get("capacity_usage"))
+            self.assertEqual(allocation_attribute_usage.value, usage)
+            self.assertEqual(
+                allocation_attribute_usage.history.first().value,
+                usage,
             )
